+++ src/components/home/OverlappingAvatars.tsx
@@ -1,4 +1,4 @@
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar.tsx";
 
 export default function OverlappingAvatars() {
     const avatars = [
