+++ src/components/home/SearchBar.tsx
@@ -1,4 +1,4 @@
-import { Input } from "@/components/ui/input";
+import { Input } from "@/components/ui/input.tsx";
 import { Search } from "lucide-react";
 
 export default function SearchBar() {
